[fix]: change function name

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -11,7 +11,7 @@
  * Return: Number of nodes
  */
 
-int print_listint(const listint_t *h)
+size_t print_listint(const listint_t *h)
 {
 	int i;
